Adding Candidate Components and Forms

+++ src/components/Company/index.tsx
@@ -3,10 +3,7 @@ import api from '../../services/api';
 import * as S from './style';
 import { FaTrashAlt, FaEdit } from 'react-icons/fa';
 import CompanyPDF from '../CompanyPDF';
-interface CompanyFormProps {
-    onSuccess: () => void;
-}
-
+import Navbar from '../Menu/Navbar';
 export interface Company {
     id: number;
     cnpj: string;
@@ -22,9 +19,7 @@ export interface Company {
 
 }
 
-const CompanyForm: React.FC<CompanyFormProps> = ({ onSuccess }) => {
-
-
+const CompanyForm: React.FC = () => {
 
     const [companies, setCompanies] = useState<Company[]>([]);
     const [editingCompanyId, setEditingCompanyId] = useState<number | null>(null);
@@ -89,7 +84,6 @@ const CompanyForm: React.FC<CompanyFormProps> = ({ onSuccess }) => {
             setNumberStreet('');
             setNeighborhood('');
 
-            onSuccess();
             loadCompanies();
             setEditingCompanyId(null);
             setIsEditing(false);
@@ -98,6 +92,9 @@ const CompanyForm: React.FC<CompanyFormProps> = ({ onSuccess }) => {
         }
     };
     const handleDeleteCompany = async (companyId: number) => {
+        const confirmDelete = window.confirm('Tem certeza que deseja excluir esta empresa?');
+        if (!confirmDelete) return;
+
         try {
             await api.delete(`/company/${companyId}`);
             loadCompanies();
@@ -127,51 +124,56 @@ const CompanyForm: React.FC<CompanyFormProps> = ({ onSuccess }) => {
     };
 
     return (
-        <S.Container>
-            <S.FormContainer onSubmit={handleSubmit}>
-                <S.InputField type="text" name="cnpj" value={cnpj} onChange={(e) => setCnpj(e.target.value)} placeholder="CNPJ" required />
-                <S.InputField type="text" name="fantasyName" value={fantasyName} onChange={(e) => setFantasyName(e.target.value)} placeholder="Nome Fantasia" required />
-                <S.InputField type="text" name="corporateReason" value={corporateReason} onChange={(e) => setCorporateReason(e.target.value)} placeholder="Razão Social" required />
-                <S.InputField type="text" name="ie" value={ie} onChange={(e) => setIe(e.target.value)} placeholder="Inscrição Estadual" required />
-                <S.InputField type="text" name="phone" value={phone} onChange={(e) => setPhone(e.target.value)} placeholder="Telefone" required />
-                <S.InputField type="text" name="street" value={street} onChange={(e) => setStreet(e.target.value)} placeholder="Rua" required />
-                <S.InputField type="text" name="numberStreet" value={numberStreet} onChange={(e) => setNumberStreet(e.target.value)} placeholder="Número" required />
-                <S.InputField type="text" name="neighborhood" value={neighborhood} onChange={(e) => setNeighborhood(e.target.value)} placeholder="Bairro" required />
-                {error && <S.Error>{error}</S.Error>}
-                <S.SubmitButton type="submit">{isEditing ? 'Editar Empresa' : 'Cadastrar Empresa'}</S.SubmitButton>
-                <CompanyPDF companies={companies} />
-            </S.FormContainer>
+        <>
+            <Navbar></Navbar>
+            <S.Title>Gerenciar Empresas</S.Title>
+            <S.Container>
 
-            <S.CompanyList>
-                <table>
-                    <thead>
-                        <tr>
-                            <S.TableHeader>CNPJ</S.TableHeader>
-                            <S.TableHeader>IE</S.TableHeader>
-                            <S.TableHeader>Razão Social</S.TableHeader>
-                            <S.TableHeader>Nome Fantasia</S.TableHeader>
-                            <S.TableHeader>Telefone</S.TableHeader>
-                            <S.TableHeader>Ações</S.TableHeader>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {companies.map((company: Company, index: number) => (
-                            <tr key={company.id} style={{ backgroundColor: index % 2 === 0 ? '#f4f4f4' : 'white' }}>
-                                <td>{company.cnpj}</td>
-                                <td>{company.ie}</td>
-                                <td>{company.corporateReason}</td>
-                                <td>{company.fantasyName}</td>
-                                <td>{company.phone}</td>
-                                <td>
-                                    <S.ActionButton onClick={() => handleDeleteCompany(company.id)}><FaTrashAlt /></S.ActionButton>
-                                    <S.ActionButton onClick={() => handleEditCompany(company.id)}><FaEdit /></S.ActionButton>
-                                </td>
+                <S.FormContainer onSubmit={handleSubmit}>
+                    <S.InputField type="text" name="cnpj" value={cnpj} onChange={(e) => setCnpj(e.target.value)} placeholder="CNPJ" required />
+                    <S.InputField type="text" name="fantasyName" value={fantasyName} onChange={(e) => setFantasyName(e.target.value)} placeholder="Nome Fantasia" required />
+                    <S.InputField type="text" name="corporateReason" value={corporateReason} onChange={(e) => setCorporateReason(e.target.value)} placeholder="Razão Social" required />
+                    <S.InputField type="text" name="ie" value={ie} onChange={(e) => setIe(e.target.value)} placeholder="Inscrição Estadual" required />
+                    <S.InputField type="text" name="phone" value={phone} onChange={(e) => setPhone(e.target.value)} placeholder="Telefone" required />
+                    <S.InputField type="text" name="street" value={street} onChange={(e) => setStreet(e.target.value)} placeholder="Rua" required />
+                    <S.InputField type="text" name="numberStreet" value={numberStreet} onChange={(e) => setNumberStreet(e.target.value)} placeholder="Número" required />
+                    <S.InputField type="text" name="neighborhood" value={neighborhood} onChange={(e) => setNeighborhood(e.target.value)} placeholder="Bairro" required />
+                    {error && <S.Error>{error}</S.Error>}
+                    <S.SubmitButton type="submit">{isEditing ? 'Editar Empresa' : 'Cadastrar Empresa'}</S.SubmitButton>
+                    <CompanyPDF companies={companies} />
+                </S.FormContainer>
+
+                <S.CompanyList>
+                    <table>
+                        <thead>
+                            <tr>
+                                <S.TableHeader>CNPJ</S.TableHeader>
+                                <S.TableHeader>IE</S.TableHeader>
+                                <S.TableHeader>Razão Social</S.TableHeader>
+                                <S.TableHeader>Nome Fantasia</S.TableHeader>
+                                <S.TableHeader>Telefone</S.TableHeader>
+                                <S.TableHeader>Ações</S.TableHeader>
                             </tr>
-                        ))}
-                    </tbody>
-                </table>
-            </S.CompanyList>
-        </S.Container>
+                        </thead>
+                        <tbody>
+                            {companies.map((company: Company, index: number) => (
+                                <tr key={company.id} style={{ backgroundColor: index % 2 === 0 ? '#f4f4f4' : 'white' }}>
+                                    <td>{company.cnpj}</td>
+                                    <td>{company.ie}</td>
+                                    <td>{company.corporateReason}</td>
+                                    <td>{company.fantasyName}</td>
+                                    <td>{company.phone}</td>
+                                    <td>
+                                        <S.ActionButton onClick={() => handleDeleteCompany(company.id)}><FaTrashAlt /></S.ActionButton>
+                                        <S.ActionButton onClick={() => handleEditCompany(company.id)}><FaEdit /></S.ActionButton>
+                                    </td>
+                                </tr>
+                            ))}
+                        </tbody>
+                    </table>
+                </S.CompanyList>
+            </S.Container>
+        </>
     );
 };
